refactor: auth mock 참조 경로 정리
